+++ lib/app.dart
@@ -1,16 +1,23 @@
-import 'package:dream/screen/intro_screen.dart';
+import 'package:dream/screen/introlduction/intro_provider.dart';
+import 'package:dream/screen/introlduction/intro_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
+    return MultiProvider(
+      providers: [
+        ChangeNotifierProvider(create: (context) => IntroProvider()),
+      ],
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.lightBlue,
+        ),
+        home: IntroScreen(),
       ),
-      home: IntroScreen(),
     );
   }
 }
