+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
   webview_flutter: ^2.0.13
   image_picker: ^0.8.4+1
   path_provider: ^2.0.1
+  provider: ^6.0.1
 dev_dependencies:
   flutter_test:
     sdk: flutter
